--- url-shortner-api/src/main/java/com/smilegate/bio/URLShortnerApplication.java
@@ -1,13 +1,0 @@
-package com.smilegate.bio;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class URLShortnerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(URLShortnerApplication.class, args);
-    }
-
-}
