--- url-shortner-api/src/main/java/com/smilegate/bio/config/JpaAuditingConfiguration.java
@@ -1,9 +1,0 @@
-package com.smilegate.bio.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfiguration {
-}
