--- url-shortner-api/src/main/java/com/smilegate/bio/config/WebConfiguration.java
@@ -1,17 +1,0 @@
-package com.smilegate.bio.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebConfiguration implements WebMvcConfigurer {
-    private static final Long MAX_AGE = 3000L;
-
-    @Override
-    public void addCorsMappings(final CorsRegistry registry) {
-        registry.addMapping("/**")
-                .allowedOrigins("*")
-                .maxAge(MAX_AGE);
-    }
-}
