--- url-shortner-api/src/main/java/com/smilegate/bio/dto/CreateShortURLRequestDTO.java
@@ -1,4 +1,0 @@
-package com.smilegate.bio.dto;
-
-public record CreateShortURLRequestDTO(String originUrl) {
-}
