--- url-shortner-api/src/main/java/com/smilegate/bio/repository/ShortURLRepository.java
@@ -1,12 +1,0 @@
-package com.smilegate.bio.repository;
-
-import com.smilegate.bio.entity.ShortURL;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ShortURLRepository extends JpaRepository<ShortURL, Long> {
-    Optional<ShortURL> findByOriginUrl(String originUrl);
-    Optional<ShortURL> findByShortUrl(String shortUrl);
-
-    Optional<ShortURL> findTopByOrderByIdDesc();
-}
