+++ url-shortner-api/src/main/java/com/smilegate/urlshortner/URLShortnerApplication.java
@@ -1,0 +1,13 @@
+package com.smilegate.urlshortner;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class URLShortnerApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(URLShortnerApplication.class, args);
+    }
+
+}
