+++ url-shortner-api/src/main/java/com/smilegate/urlshortner/config/JpaAuditingConfiguration.java
@@ -1,0 +1,9 @@
+package com.smilegate.urlshortner.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfiguration {
+}
