+++ url-shortner-api/src/main/java/com/smilegate/urlshortner/config/WebConfiguration.java
@@ -1,0 +1,17 @@
+package com.smilegate.urlshortner.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfiguration implements WebMvcConfigurer {
+    private static final Long MAX_AGE = 3000L;
+
+    @Override
+    public void addCorsMappings(final CorsRegistry registry) {
+        registry.addMapping("/**")
+                .allowedOrigins("*")
+                .maxAge(MAX_AGE);
+    }
+}
