+++ url-shortner-api/src/main/java/com/smilegate/urlshortner/dto/CreateShortURLRequestDTO.java
@@ -1,0 +1,4 @@
+package com.smilegate.urlshortner.dto;
+
+public record CreateShortURLRequestDTO(String originUrl) {
+}
