--- url-shortner-api/src/test/java/com/smilegate/bio/URLShortnerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.smilegate.bio;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class URLShortnerApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
