+++ url-shortner-api/src/test/java/com/smilegate/urlshortner/URLShortnerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.smilegate.urlshortner;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class URLShortnerApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
